+++ payments_payu/provider.py
@@ -689,7 +689,11 @@ class PayuProvider(BasicProvider):
             )
         if refund_status == "CANCELED":
             raise ValueError(f"refund {refund_id} of payment {payment.id} canceled")
-        elif refund_status not in {"PENDING", "FINALIZED"}:
+        elif refund_status == "FINALIZED":
+            raise NotImplementedError(
+                f"refund {refund_id} of payment {payment.id} being FINALIZED already is not supported yet"
+            )
+        elif refund_status not in {"PENDING"}:
             raise PayuApiError(
                 f"invalid status of refund {refund_id} of payment {payment.id}"
             )
@@ -698,7 +702,14 @@ class PayuProvider(BasicProvider):
                 f"refund {refund_id} of payment {payment.id} in different currency not supported yet: "
                 f"{refund_currency}"
             )
-        return refund_amount
+        if amount is not None and refund_amount != amount:
+            raise NotImplementedError(
+                f"refund {refund_id} of payment {payment.id} having a different amount than requested not supported "
+                f"yet: {refund_amount}"
+            )
+        # Return 0 in order not to change captured_amount yet. If we returned the amount, captured_amount would change
+        # twice (now and once we get a notification from PayU).
+        return Decimal(0)
 
 
 class PaymentProcessor(object):